% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advance.R
\name{advance}
\alias{advance}
\title{Advance an ODE system}
\usage{
advance(x, tspan)
}
\arguments{
\item{x}{mrgsolve model object.}

\item{tspan}{vector of output times.}
}
\description{
Advance an ODE system
}
