% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rx.R
\docType{methods}
\name{as.ev,sig-method}
\alias{as.ev,sig-method}
\title{Coerce sig object to event}
\usage{
\S4method{as.ev}{sig}(x, ...)
}
\arguments{
\item{x}{a sig object}

\item{...}{not used}
}
\description{
Coerce sig object to event
}
\keyword{internal}
