% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mrgmod.R
\docType{methods}
\name{as.list,mrgmod-method}
\alias{as.list,mrgmod-method}
\title{Coerce a model object to list.}
\usage{
\S4method{as.list}{mrgmod}(x, deep = FALSE, ...)
}
\arguments{
\item{x}{mrgmod object}

\item{deep}{if \code{TRUE}, extra information is returned
(see details).}

\item{...}{not used}
}
\description{
Coerce a model object to list.
}
\details{
If \code{deep} is \code{TRUE}, then the values for
\code{trans}, \code{advan}, and \code{mindt} are
returned as well as a summary of internal model 
functions (with a call to \code{mrgsolve:::funset}).
}
