% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modspec.R
\name{BLOCK_PARSE}
\alias{BLOCK_PARSE}
\alias{CAPTURE}
\alias{CMT}
\alias{FIXED}
\alias{INIT}
\alias{PARAM}
\alias{THETA}
\title{Functions to parse code blocks.}
\usage{
PARAM(x, env, annotated = FALSE, pos = 1, ...)

FIXED(x, env, annotated = FALSE, pos = 1, ...)

THETA(x, env, annotated = FALSE, pos = 1, name = "THETA", ...)

INIT(x, env, annotated = FALSE, pos = 1, ...)

CMT(x, env, annotated = FALSE, pos = 1, ...)

CAPTURE(x, env, annotated = FALSE, pos = 1, ...)
}
\arguments{
\item{x}{data}

\item{env}{parse environment}

\item{annotated}{logical}

\item{pos}{block position}

\item{...}{passed}

\item{name}{block name}
}
\description{
Most of the basic blocks are listed in this help topic.  
But see also \code{\link{PKMODEL}} which has more-involved 
options and is documented separately.
}
\seealso{
\code{\link{PKMODEL}}
}

