% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/class_mrgmod.R
\docType{methods}
\name{blocks}
\alias{blocks}
\alias{blocks,mrgmod-method}
\alias{blocks,character-method}
\title{Return the code blocks from a model specification file.}
\usage{
blocks(x, ...)

\S4method{blocks}{mrgmod}(x, ...)

\S4method{blocks}{character}(x, ...)
}
\arguments{
\item{x}{model object or path to model specification file}

\item{...}{passed along}
}
\description{
Return the code blocks from a model specification file.
}
\examples{
mod <- mrgsolve:::house()
mod \%>\% blocks
mod \%>\% blocks(PARAM,TABLE)

}
