% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{cama}
\alias{cama}
\title{Run the model cama function.}
\usage{
cama(mod, fn = "cama", ...)
}
\arguments{
\item{mod}{model object}

\item{fn}{function name}

\item{...}{passed to update}
}
\description{
Run the model cama function.
}
\details{
\code{sah-mah}
}
