% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modspec.R
\name{CAPTURE}
\alias{CAPTURE}
\title{Parse \code{$CAPTURE} block.}
\usage{
CAPTURE(x, env, annotated = FALSE, pos = 1, ...)
}
\arguments{
\item{x}{block text}

\item{env}{parse environment}

\item{annotated}{logical}

\item{pos}{parse position}

\item{...}{not used}
}
\description{
Parse \code{$CAPTURE} block.
}

