% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{chain}
\alias{chain}
\title{Functions for chaining commands together.}
\description{
Use these functions with chaining commands togehter with the %>%
operator.
}
\details{
Other functions that may be used in the chain of commands include: 
\code{\link{param}}, \code{\link{init}}, \code{\link[mrgsolve]{update}},
\code{\link{ev}}.
or any other function that will take the output of the preceeding command as it's first argument.
}
\examples{

mod <- mrgsolve:::house()

data(exidata)
data(exTheoph)

out <- mod \%>\% data_set(exTheoph) \%>\% mrgsim()
out <- mod \%>\% carry_out(evid) \%>\% ev(amt=100, cmt=1) \%>\% mrgsim()
out <- mod \%>\% Req(CP,RESP) \%>\% mrgsim()
}
