% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inven.R
\name{check_data_names}
\alias{check_data_names}
\title{Check data set against model parameters}
\usage{
check_data_names(
  obj,
  x,
  covariates = TRUE,
  strict = FALSE,
  silent = FALSE,
  tags = NULL
)
}
\arguments{
\item{obj}{a data frame or other object with names to check.}

\item{x}{a model object.}
}
\description{
Check data set against model parameters
}
