% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/class_mrgmod.R
\docType{methods}
\name{cmtn}
\alias{cmtn}
\alias{cmtn,mrgmod-method}
\title{Get the compartment number from a compartment name.}
\usage{
cmtn(x, ...)

\S4method{cmtn}{mrgmod}(x, tag, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed along}

\item{tag}{compartment name}
}
\description{
Get the compartment number from a compartment name.
}
\examples{
mod <- mrgsolve:::house()
mod \%>\% cmtn("CENT")
}
