% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modlib.R
\name{code}
\alias{code}
\title{Extract the code from a model}
\usage{
code(x)
}
\arguments{
\item{x}{an mrgsolve model object}
}
\value{
A character vector of model code.
}
\description{
This function is currently not exported, so be sure to call it with
\code{mrgsolve:::code(...)}.
}
\examples{
mod <- mrgsolve::house()
mrgsolve:::code(mod)

}
