% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modspec.R
\name{convert_fortran_if}
\alias{convert_fortran_if}
\title{Convert fortran IF ELSE THEN}
\usage{
convert_fortran_if(x)
}
\arguments{
\item{x}{string to convert}
}
\description{
Convert fortran IF ELSE THEN
}
\examples{
convert_fortran_if("IF ( WT .GE.90) THEN")

}
