% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modspec.R
\name{convert_pow}
\alias{convert_pow}
\title{Convert power from fortran to C++ style}
\usage{
convert_pow(x)
}
\arguments{
\item{x}{string to convert.}
}
\description{
Convert power from fortran to C++ style
}
\examples{
convert_pow("a**2")
convert_pow("(WT/70) ** THETA(3)")

}
