% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modspec.R
\name{convert_semicolons}
\alias{convert_semicolons}
\title{Add semicolons to code blocks}
\usage{
convert_semicolons(x)
}
\arguments{
\item{x}{string to convert.}
}
\description{
Add semicolons to code blocks
}
\examples{
convert_semicolons("CL = THETA1")

}
