% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmutate.R
\docType{methods}
\name{covset}
\alias{as.covset}
\alias{as.list,covset-method}
\alias{covset}
\title{Create a set of covariates.}
\usage{
covset(...)

\S4method{as.list}{covset}(x, ...)

as.covset(x)
}
\arguments{
\item{...}{formulae to use for the covset}

\item{x}{a covset}
}
\description{
Create a set of covariates.
}

