% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{custom_tol}
\alias{custom_tol}
\alias{customize_rtol}
\alias{customize_atol}
\title{Customize tolerances for specific compartments}
\usage{
customize_rtol(.x, .rtol = list(), .default = NULL, .use = TRUE, ...)

customize_atol(.x, .atol = list(), .default = NULL, .use = TRUE, ...)
}
\arguments{
\item{.x}{a model object.}

\item{.rtol}{a named numeric list or vector, where names reference
selected model compartments and relative tolerances for those
compartments.}

\item{.default}{the default tolerance value to use for compartments
not listed in \code{.rtol}, \code{.atol}, or \code{...}; if not supplied, the current
scalar value in \code{.x} will be used.}

\item{.use}{\code{logical}; if \code{TRUE}, then a call to \code{\link[=use_custom_tol]{use_custom_tol()}} will
be made prior to return; otherwise, a call to \code{\link[=use_scalar_tol]{use_scalar_tol()}} will
be made; under expected use, the value for this argument is kept \code{TRUE},
so that whenever tolerances are customized, they will be used in the
next simulation run.}

\item{.atol}{a named numeric list or vector, where names reference
selected model compartments and absolute tolerances for those
compartments.}
}
\description{
These functions update the relative or absolute tolerance values
only for the custom tolerance configuration.
}
