% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{customize_rtol}
\alias{customize_rtol}
\alias{customize_atol}
\alias{reset_rtol}
\alias{reset_atol}
\title{Customize tolerances for specific compartments}
\usage{
customize_rtol(.x, .rtol = list(), .default = NULL, ...)

customize_atol(.x, .atol, .default = NULL, ...)

reset_rtol(x, default = NULL)

reset_atol(x, default = NULL)
}
\description{
Customize tolerances for specific compartments
}
