% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{cvec}
\alias{cvec}
\alias{cvec,character-method}
\alias{s_}
\title{Create create character vectors}
\usage{
cvec(x, ...)

\S4method{cvec}{character}(x)

s_(...)
}
\arguments{
\item{x}{comma-separated quoted string (for \code{cvec})}

\item{...}{unquoted strings (for \code{ch})}
}
\description{
Create create character vectors
}
\examples{

cvec("A,B,C")
s_(A,B,C)

}
