% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsim.R
\name{data_qsim}
\alias{data_qsim}
\title{DEPRECATED: Create a matrix of events and observations for simulation}
\usage{
data_qsim(e, times)
}
\arguments{
\item{e}{an event object}

\item{times}{numeric vector of observation times or a 
\code{tgrid} object}
}
\description{
This function is to be used with \code{\link{qsim}} only.
}
\keyword{internal}
