% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.R
\name{details}
\alias{details}
\title{Extract model details.}
\usage{
details(x, complete = FALSE, values = FALSE, ...)
}
\arguments{
\item{x}{a model object}

\item{complete}{logical; if \code{TRUE}, un-annotated parameters and compartments will be added to the output}

\item{values}{logical; if \code{TRUE}, a values column will be added to the output}

\item{...}{not used}
}
\description{
Extract model details.
}
