% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmutate.R
\name{dmutate}
\alias{dmutate}
\alias{dmutate_}
\title{Add random variates to a data frame.}
\usage{
dmutate(data, form, ...)

dmutate_(data, x, ...)
}
\arguments{
\item{data}{the data.frame to mutate}

\item{form}{an unquoted R formula; see details}

\item{...}{additional inputs}

\item{x}{formula}
}
\description{
Add random variates to a data frame.
}

