% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve.R
\name{do_mrgsim}
\alias{do_mrgsim}
\title{Do an mrgsim run}
\usage{
do_mrgsim(x, data, idata = null_idata, carry.out = character(0),
  seed = as.integer(NA), Request = character(0), capture = NULL,
  obsonly = FALSE, obsaug = FALSE, tgrid = NULL, recsort = 1,
  deslist = list(), descol = character(0), filbak = TRUE, tad = FALSE,
  nocb = TRUE, skip_init_calc = FALSE, ...)
}
\arguments{
\item{x}{a model object}

\item{data}{a simulation data set}

\item{idata}{individual-level data}

\item{carry.out}{data items to copy into the output}

\item{seed}{deprecated}

\item{Request}{compartments or captured variables to retain
in the simulated output}

\item{capture}{character file name used for debugging (not related
to \code{$CAPTURE}}

\item{obsonly}{if \code{TRUE}, dosing records are not included
in the output}

\item{obsaug}{augment the data set with time grid observations}

\item{tgrid}{a tgrid object; or a numeric vector of simulation times
or another object with an \code{stime} method}

\item{recsort}{record sorting flag}

\item{deslist}{a list of tgrid objects}

\item{descol}{the name of a column for assigning designs}

\item{filbak}{carry data items backward when the first 
data set row has time greater than zero}

\item{tad}{if \code{TRUE}, include time after dose
in the output}

\item{nocb}{if \code{TRUE}, use next observation carry 
backward method; otherwise, use \code{locf}}

\item{skip_init_calc}{don't use \code{$MAIN} to 
calculate initial conditions}

\item{...}{not used}
}
\description{
Do an mrgsim run
}
