% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_assign}
\alias{env_assign}
\title{Add objects to model environment}
\usage{
env_assign(.x, x, value)
}
\arguments{
\item{.x}{model object}

\item{x}{passed to \code{\link{assign}}}

\item{value}{passed to \code{\link{assign}}}
}
\description{
Add objects to model environment
}
