% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_eval}
\alias{env_eval}
\title{Re-evaluate the code in the ENV block.}
\usage{
env_eval(x, seed = NULL)
}
\arguments{
\item{x}{model object}

\item{seed}{passed to \code{\link{set.seed}} if a numeric value is supplied}
}
\description{
The \code{$ENV} block is a block of R code that can 
realize any sort of R object that might be used in running 
a model.
}
\seealso{
\code{\link{env_get}}, \code{\link{env_ls}}
}
