% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_update}
\alias{env_update}
\title{Update objects in model environment.}
\usage{
env_update(.x, ..., .dots = list())
}
\arguments{
\item{.x}{model object}

\item{...}{objects to update}

\item{.dots}{list of objects to updated}
}
\description{
Update objects in model environment.
}
