% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{eta_source}
\alias{eta_source}
\title{Manipulate how ETA are generated}
\usage{
eta_source(x, location = c("omega", "data", "data.all"))
}
\arguments{
\item{x}{model object.}

\item{location}{the source for ETA in the simulation; see details.}
}
\value{
The updated model object is returned.
}
\description{
Manipulate how ETA are generated
}
