% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{etas_from}
\alias{etas_from}
\title{Manipulate how ETA are generated}
\usage{
etas_from(x, source = c("omega", "data", "data.all"))
}
\arguments{
\item{x}{model object.}

\item{source}{the source for ETA in the simulation; see details.}
}
\value{
The updated model object is returned.
}
\description{
Manipulate how ETA are generated
}
