% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ev.R
\name{$,ev-method}
\alias{$,ev-method}
\alias{[[,ev-method}
\title{Select columns from an ev object}
\usage{
\S4method{$}{ev}(x, name)

\S4method{[[}{ev}(x, i, exact = TRUE)
}
\arguments{
\item{x}{ev object}

\item{name}{column to select}

\item{i}{an element to select}

\item{exact}{not used}
}
\description{
Select columns from an ev object
}
