% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\docType{methods}
\name{ev_ops}
\alias{ev_ops}
\alias{+,ev,ev-method}
\alias{\%then\%}
\alias{\%then\%,ev,ev-method}
\alias{+,ev,numeric-method}
\alias{c,ev-method}
\title{Operations for ev objects.}
\usage{
\S4method{+}{ev,ev}(e1, e2)

e1 \%then\% e2

\S4method{\%then\%}{ev,ev}(e1, e2)

\S4method{+}{ev,numeric}(e1, e2)

\S4method{c}{ev}(x, ..., recursive = TRUE)
}
\arguments{
\item{e1}{object on left hand side of operator (lhs)}

\item{e2}{object on right hand side of operator (rhs)}

\item{x}{an ev object}

\item{...}{other ev objects to collect}

\item{recursive}{not used}
}
\description{
Operations for ev objects.
}
\details{
All operations involving \code{\link[=mrgmod-class]{mrgmod}} 
objects have been deprecated.
}
