% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{ev_rep}
\alias{ev_rep}
\title{Replicate an event object}
\usage{
ev_rep(x, id, as.ev = FALSE)
}
\arguments{
\item{x}{event object}

\item{id}{numeric vector if IDs}

\item{as.ev}{if \code{TRUE} an event object is returned}
}
\value{
A single event object or event object as 
determined by the value of \code{as.ev}.
}
\description{
Replicate an event object
}
\examples{

e1 <- c(ev(amt=100), ev(amt=200, ii=24, addl=2, time=72))

ev_rep(e1, 1:5)

}
