% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rx.R
\docType{methods}
\name{ev_rx}
\alias{ev_rx}
\alias{ev_rx,mrgmod,character-method}
\alias{ev_rx,character,ANY-method}
\title{Create interventions from Rx input}
\usage{
ev_rx(x, y, ...)

\S4method{ev_rx}{mrgmod,character}(x, y, ...)

\S4method{ev_rx}{character,ANY}(x, df = FALSE, ...)
}
\arguments{
\item{x}{a model object}

\item{y}{Rx input; see \link{parse_rx} for details}

\item{...}{not used at this time}

\item{df}{if \code{TRUE} then a data frame is returned}
}
\description{
See \link{parse_rx} for Rx specification.
}
\examples{

ev_rx("100 q 12 x 3")
}
\seealso{
\link{parse_rx}
}
