% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_evdata.R
\name{evd,missing-method}
\alias{evd,missing-method}
\title{Create an evdata object}
\usage{
\S4method{evd}{missing}(x, ...)
}
\arguments{
\item{...}{passed to \code{\link[=ev]{ev()}}}
}
\description{
Create an evdata object
}
