% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mrgmod.R
\docType{methods}
\name{events}
\alias{events}
\alias{events,mrgmod-method}
\title{DEPRECATED: get an events object from a model object}
\usage{
events(x, ...)

\S4method{events}{mrgmod}(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{not used}
}
\description{
DEPRECATED: get an events object from a model object
}
