% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mrgmod.R
\name{file_show}
\alias{file_show}
\title{Show model specification and C++ files.}
\usage{
file_show(x, spec = TRUE, source = TRUE, ...)
}
\arguments{
\item{x}{model object}

\item{spec}{logical; show the model specification file}

\item{source}{logical; show the C++ file that is actually compiled}

\item{...}{not used}
}
\description{
Show model specification and C++ files.
}
