% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mrgmod.R
\name{get_env}
\alias{get_env}
\title{Return model \code{$ENV}.}
\usage{
get_env(x, tolist = TRUE)
}
\arguments{
\item{x}{model object}

\item{tolist}{should the environment be coreced to \code{list}?}
}
\description{
Return model \code{$ENV}.
}

