% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-tol.R
\name{get_tol}
\alias{get_tol}
\alias{get_tol_list}
\title{Extract rtol and atol information from a model object}
\usage{
get_tol(x)

get_tol_list(x)
}
\arguments{
\item{x}{a model object.}
}
\value{
A data frame (\code{get_tol()}) or a named list (\code{get_tol_list()}).
}
\description{
Extract rtol and atol information from a model object
}
\examples{
mod <- house()
get_tol(mod)
get_tol_list(mod)

}
\seealso{
\code{\link[=reset_tol]{reset_tol()}}, \code{\link[=custom_tol]{custom_tol()}}, \code{\link[=use_custom_tol]{use_custom_tol()}}, \code{\link[=use_scalar_tol]{use_scalar_tol()}}
}
