% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modspec.R
\name{PARAM}
\alias{PARAM}
\title{Parse \code{$PARAM} block.}
\usage{
PARAM(x, env, annotated = FALSE, pos = 1, ...)
}
\arguments{
\item{x}{data}

\item{env}{parse environment}

\item{annotated}{logical}

\item{pos}{block position}

\item{...}{passed}
}
\description{
Parse \code{$PARAM} block.
}

