% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modspec.R
\name{THETA}
\alias{THETA}
\title{Parse \code{$THETA} block.}
\usage{
THETA(x, env, annotated = FALSE, pos = 1, name = "THETA", ...)
}
\arguments{
\item{x}{data}

\item{env}{parse environment}

\item{annotated}{logical}

\item{pos}{parse position}

\item{name}{character prefix for parameter names}

\item{...}{passed}
}
\description{
Parse \code{$THETA} block.
}

