% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mrgsims_arrow.R
\name{head,mrgsimsds-method}
\alias{head,mrgsimsds-method}
\title{Return the first several rows of the object.}
\usage{
\S4method{head}{mrgsimsds}(x, n = 6L, ...)
}
\arguments{
\item{x}{a object with class \code{mrgsimsds}.}

\item{n}{number of rows to show.}

\item{...}{passed to \code{\link[=head]{head()}}.}
}
\description{
Return the first several rows of the object.
}
