% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mrgmod.R
\name{house}
\alias{house}
\title{Return a pre-compiled, PK/PD model}
\usage{
house(...)
}
\arguments{
\item{...}{passed to \code{\link[=update]{update()}}.}
}
\value{
A \code{packmod} object, ready to simulate.
}
\description{
This model is used in a lot of package tests, but it can be useful to
support user experimentation as well.
}
\examples{

mod <- mrgsolve::house()

see(mod)

mod \%>\% ev(amt = 100) \%>\% mrgsim() \%>\% plot()

}
\keyword{internal}
