% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idata_set.R
\docType{methods}
\name{idata_set}
\alias{idata_set}
\alias{idata_set,mrgmod,ANY-method}
\alias{idata_set,mrgmod,data.frame-method}
\alias{idata_set,mrgmod,missing-method}
\title{Set the \code{idata} argument for \code{mrgsim}.}
\usage{
idata_set(x, data, ...)

\S4method{idata_set}{mrgmod,data.frame}(x, data, subset = TRUE,
  select = TRUE, covset = NULL, ...)

\S4method{idata_set}{mrgmod,ANY}(x, data, ...)

\S4method{idata_set}{mrgmod,missing}(x, object, ...)
}
\arguments{
\item{x}{model object}

\item{data}{a data set coercable to data.frame}

\item{...}{passed along}

\item{subset}{passed to \code{dplyr::filter_}}

\item{select}{passed to \code{dplyr::select_}}

\item{covset}{the name of a \code{\link{covset}} object in `$ENV`}

\item{object}{character name of an object existing in \code{$ENV} to use for the data set}
}
\description{
Set the \code{idata} argument for \code{mrgsim}.
}

