% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mrgsims.R
\name{is.mrgsims}
\alias{is.mrgsims}
\title{Check if an object is mrgsim output.}
\usage{
is.mrgsims(x)
}
\arguments{
\item{x}{any object}
}
\value{
\code{TRUE} if \code{x} inherits \code{mrgsims}.
}
\description{
Check if an object is mrgsim output.
}
