% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mrgmod.R
\name{ls_env}
\alias{ls_env}
\title{List objects in the model environment.}
\usage{
ls_env(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed to \code{\link{ls}}}
}
\description{
Each model keeps an internal environment that allows the user 
to carry any \code{R} object along.  Objects are coded in \code{$ENV}.
}

