% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlist.R
\docType{methods}
\name{c,matlist-method}
\alias{c,matlist-method}
\title{Operations with matlist objects.}
\usage{
\S4method{c}{matlist}(x, ..., recursive = FALSE)
}
\arguments{
\item{x}{a matlist object}

\item{...}{other matlist objects}

\item{recursive}{not used}
}
\description{
Operations with matlist objects.
}
