% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge.list}
\alias{merge.list}
\title{Merge two lists.}
\usage{
\method{merge}{list}(x, y, ..., open = FALSE, warn = TRUE,
  context = "object", wild = "...")
}
\arguments{
\item{x}{the original list}

\item{y}{the new list for merging}

\item{...}{not used}

\item{open}{logical indicating whether or not new items should be allowed in the list upon merging.}

\item{warn}{issue warning if nothing found to update}

\item{context}{description of usage context}

\item{wild}{wild-card name; see details}
}
\description{
Merge two lists.
}
\details{
Wild-card names (\code{wild}) are always retained in \code{x} and are brought along from \code{y} only when \code{open}.
}
