% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mrgsims.R
\docType{methods}
\name{mod}
\alias{mod}
\alias{mod,mrgsims-method}
\title{Return the model object.}
\usage{
mod(x, ...)

\S4method{mod}{mrgsims}(x, ...)
}
\arguments{
\item{x}{mrgsims object}

\item{...}{passed along}
}
\description{
Return the model object.
}
