% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modlib.R
\name{modlib_pkpd}
\alias{modlib_pkpd}
\title{modlib: Pharmacokinetic / pharmacodynamic models}
\description{
modlib: Pharmacokinetic / pharmacodynamic models
}
\details{
See \link{modlib_details} for more detailed descriptions of parameters and
compartments.

All PK/PD models include 2-compartment PK model with absorption from
2 extravascular compartments and linear + nonlinear clearance. The
PK models are parameterized with \code{CL}, \code{V2}, \code{Q}, \verb{V3, }VMAX\verb{, }KM\verb{, }KA\code{and}KA2\verb{and implement compartments}EV\verb{, }EV2\verb{, }CENT\verb{, }PERIPH\verb{. The indirect response models have compartment }RESP\verb{and the emax model has output variable}RESP\verb{. PD parameters include }KIN\verb{, }KOUT\verb{, }IC50\verb{, }EC50\verb{, }IMAX\verb{, }EMAX\verb{, }E0\verb{, and }n`.

Also, once the model is loaded, use the \code{\link[=see]{see()}} method for \code{mrgmod} to
view the model code.
}
\section{Model description}{

\itemize{
\item \code{irm1}: inhibition of response production
\item \code{irm2}: inhibition of response loss
\item \code{irm3}: stimulation of response production
\item \code{irm4}: stimulation of response loss
\item \code{pd_effect}: effect compartment model
\item \code{emax}: sigmoid emax model
}
}

