% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_modlist.R
\docType{class}
\name{modlist-class}
\alias{modlist-class}
\alias{$,modlist-method}
\title{S4 class matlist.}
\usage{
\S4method{$}{modlist}(x, name)
}
\arguments{
\item{x}{modlist object}

\item{name}{model to take; used with \code{$}}
}
\description{
S4 class matlist.
}
