% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwrite.R
\name{mread_yaml}
\alias{mread_yaml}
\alias{mread_json}
\title{Read a model from yaml or json format}
\usage{
mread_yaml(file, model = basename(file), project = tempdir(), ...)

mread_json(file, model = basename(file), project = tempdir(), ...)
}
\arguments{
\item{file}{the yaml or json file name.}

\item{model}{a new model name to use when calling `mread_yaml` or 
`mread_json`.}

\item{project}{the directory where the model should be built.}

\item{...}{passed to [mread()].}
}
\value{
A model object.
}
\description{
Read a model from yaml or json format
}
