% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mrgmod.R
\docType{methods}
\name{$,mrgmod-method}
\alias{$,mrgmod-method}
\alias{[[,mrgmod-method}
\alias{[,mrgmod-method}
\title{Select parameter values from a model object}
\usage{
\S4method{$}{mrgmod}(x, name)

\S4method{[[}{mrgmod}(x, i, exact = TRUE)

\S4method{[}{mrgmod}(x, i)
}
\arguments{
\item{x}{mrgmod object}

\item{name}{parameter to take}

\item{i}{an element to select}

\item{exact}{not used}
}
\description{
The \code{$} and \code{[[} operators get the value 
of a single parameter in the model.  The 
\code{[} gets several values, returning a 
named list.
}
