% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsim_dq.R
\name{mrgsim_dq}
\alias{mrgsim_dq}
\title{Do a simplified, quicker simulation run}
\usage{
mrgsim_dq(x, data, recsort = 1, skip_init_calc = FALSE, ...)
}
\arguments{
\item{x}{a model object}

\item{data}{a simulation data set}

\item{recsort}{record sorting flag}

\item{skip_init_calc}{don't use \code{$MAIN} to 
calculate initial conditions}

\item{...}{passed to update}
}
\description{
Do a simplified, quicker simulation run
}
