% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mrgsims.R
\docType{class}
\name{mrgsims-class}
\alias{mrgsims-class}
\title{S4 class for mrgsolve simulation output.}
\description{
S4 class for mrgsolve simulation output.
}
\section{Slots}{

\describe{
\item{\code{request}}{character vector of compartments requested in simulated output}

\item{\code{outnames}}{character vector of column names in simulated output coming from table step}

\item{\code{data}}{matrix of simulated data}

\item{\code{mod}}{the mrgmod model object}
}}

