% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{mrgsolve_example}
\alias{mrgsolve_example}
\title{Extract example model from system library}
\usage{
mrgsolve_example(model = c("pkExample", "pkpdExample", "firstmodeExample",
  "viralExample", "popExample"), project = getwd(), overwrite = FALSE,
  quiet = FALSE, ...)
}
\arguments{
\item{model}{name of model}

\item{project}{working directory}

\item{overwrite}{passed to file.copy}

\item{quiet}{don't print any status messages to the screen}

\item{...}{additional arguments}
}
\description{
Extract example model from system library
}
