% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{mrgsolve_template}
\alias{mrgsolve_template}
\title{Create model specification file from template}
\usage{
mrgsolve_template(model = "template", project = getwd(),
  writeable = FALSE, overwrite = FALSE)
}
\arguments{
\item{model}{name of the model to create}

\item{project}{working directory}

\item{writeable}{logical; if \code{TRUE}, parameters may be overwritten in the  \code{main} block}

\item{overwrite}{logical; if \code{TRUE}, an existing file with same stem will be overwritten}
}
\description{
Create model specification file from template
}
