% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ev.R
\name{mutate.ev}
\alias{mutate.ev}
\title{Mutate an events object}
\usage{
\method{mutate}{ev}(.data, ...)
}
\arguments{
\item{.data}{the event object}

\item{...}{passed to \code{dplyr::mutate}}
}
\description{
Mutate an events object
}
