% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmutate.R
\docType{methods}
\name{mutate_random}
\alias{mutate_random}
\alias{mutate_random,data.frame,character-method}
\alias{mutate_random,data.frame,covset-method}
\alias{mutate_random,data.frame,formula-method}
\alias{mutate_random,data.frame,list-method}
\title{Add random variates to a data frame.}
\usage{
mutate_random(data, input, ...)

\S4method{mutate_random}{data.frame,formula}(data, input, ...)

\S4method{mutate_random}{data.frame,character}(data, input, ...)

\S4method{mutate_random}{data.frame,list}(data, input, ...)

\S4method{mutate_random}{data.frame,covset}(data, input, ...)
}
\arguments{
\item{data}{the data.frame to mutate}

\item{input}{an unquoted R formula; see details.}

\item{...}{additional inputs}
}
\description{
Add random variates to a data frame.
}

