% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mvgauss}
\alias{mvgauss}
\title{Simulate from a multivariate normal distribution with mean zero.}
\usage{
mvgauss(mat, n = 10, seed = NULL)
}
\arguments{
\item{mat}{a positive-definite matrix}

\item{n}{number of variates to simulate}

\item{seed}{if not null, passed to set.seed}
}
\description{
Simulate from a multivariate normal distribution with mean zero.
}
