% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport.R
\name{mwrite_yaml}
\alias{mwrite_yaml}
\alias{mwrite_json}
\title{Write model code to yaml or json format}
\usage{
mwrite_yaml(x, file = NULL, digits = 8)

mwrite_json(x, file = NULL, digits = 8)
}
\arguments{
\item{x}{a model object.}

\item{file}{output file name.}

\item{digits}{precision to use when writing outputs.}
}
\value{
A list containing data that was written out to the yaml or json file.
}
\description{
Write model code to yaml or json format
}
\examples{
mod <- house()

temp1 <- tempfile(fileext = ".yaml")

x <- mwrite_yaml(mod, temp1)

readLines(temp1)

temp2 <- tempfile(fileext = ".json")

y <- mwrite_json(mod, temp2)

code <- readLines(temp2) 

}
