% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mrgmod.R
\name{names,mrgmod-method}
\alias{names,mrgmod-method}
\title{Get all names from a model object}
\usage{
\S4method{names}{mrgmod}(x)
}
\arguments{
\item{x}{the model object}
}
\description{
Get all names from a model object
}
\examples{
mod <- mrgsolve::house()
names(mod)

}
