% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmxml.R
\name{nmext}
\alias{nmext}
\title{Import model estimates from a NONMEM ext file}
\usage{
nmext(
  file,
  theta = TRUE,
  omega = TRUE,
  sigma = TRUE,
  olabels = NULL,
  slabels = NULL,
  oprefix = "",
  sprefix = "",
  tname = "THETA",
  oname = "...",
  sname = "..."
)
}
\arguments{
\item{file}{the complete path to the \code{run.xml} file}

\item{theta}{logical; if TRUE, the \code{$THETA} vector is returned}

\item{omega}{logical; if TRUE, the \code{$OMEGA} matrix is returned}

\item{sigma}{logical; if TRUE, the \code{$SIGMA} matrix is returned}

\item{olabels}{labels for \code{$OMEGA}}

\item{slabels}{labels for \code{$SIGMA}}

\item{oprefix}{prefix for \code{$OMEGA} labels}

\item{sprefix}{prefix for \code{$SIGMA} labels}

\item{tname}{name for \code{$THETA}}

\item{oname}{name for \code{$OMEGA}}

\item{sname}{name for \code{$SIGMA}}
}
\description{
Import model estimates from a NONMEM ext file
}
\seealso{
nmxml
}
