% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_numericlist.R
\docType{class}
\name{numericlist-class}
\alias{numericlist-class}
\title{S4 class numeric list.}
\arguments{
\item{data}{list of data}

\item{pattern}{character of length 1 containing regular expression to be used as a filter when printing data to the console}
}
\description{
S4 class numeric list.
}
