% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_numericlist.R
\docType{methods}
\name{numericlist}
\alias{numericlist}
\alias{as.list,numericlist-method}
\alias{as.numeric,numericlist-method}
\alias{as.data.frame,numericlist-method}
\alias{length,numericlist-method}
\alias{names,numericlist-method}
\alias{$,numericlist-method}
\alias{[,numericlist-method}
\title{Methods for numericlist.}
\usage{
\S4method{as.list}{numericlist}(x, ...)

\S4method{as.numeric}{numericlist}(x)

\S4method{as.data.frame}{numericlist}(x, row.names = NULL, optional = FALSE,
  ...)

\S4method{length}{numericlist}(x)

\S4method{names}{numericlist}(x)

\S4method{$}{numericlist}(x, name)

\S4method{[}{numericlist}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{object}

\item{...}{passed along to other methods}

\item{row.names}{passed to \code{\link{as.data.frame}}}

\item{optional}{passed to \code{\link{as.data.frame}}}

\item{name}{column to take}

\item{i}{elements to keep}

\item{j}{not used}

\item{drop}{not used}
}
\description{
These methods can be used to corece \code{param} and \code{init} objects into common \code{R} data structures,
extract elements from \code{numericlist}s, or get attributes from \code{numericlist}s.
}
