% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{obsaug}
\alias{obsaug}
\title{Augment observations in the simulated output.}
\usage{
obsaug(x, value = TRUE, ...)
}
\arguments{
\item{x}{model object}

\item{value}{the value for \code{obsaug}}

\item{...}{passed along
There is also a \code{obsaug} argument to \code{\link{mrgsim}} that can 
be set to accomplish the same thing as a call to \code{obsaug} in 
the pipeline.}
}
\description{
Augment observations in the simulated output.
}
