% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{obsonly}
\alias{obsonly}
\title{Set the \code{obsonly} argument for \code{mrgsim}.}
\usage{
obsonly(x, value = TRUE, ...)
}
\arguments{
\item{x}{model object}

\item{value}{the value for \code{obsonly}}

\item{...}{passed along}
}
\description{
Set the \code{obsonly} argument for \code{mrgsim}.
}

