% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mrgmod.R
\name{outvars}
\alias{outvars}
\title{Show names of current output variables}
\usage{
outvars(x, unlist = FALSE)
}
\arguments{
\item{x}{mrgmod object}

\item{unlist}{if `TRUE` then a character vector (rather than list) is 
returned}
}
\value{
When `unlist` is `FALSE` (default) : a named list, with `cmt` showing names 
of output compartments and `capture` giving names of output variables in 
capture.  When `unlist` is `TRUE`, then a single, unnamed character vector
of outvar names is returned.
}
\description{
Show names of current output variables
}
\examples{

outvars(mrgsolve:::house())

}
