% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inven.R
\name{param_tags}
\alias{param_tags}
\title{Return parameter tags}
\usage{
param_tags(x)
}
\arguments{
\item{x}{mrgsolve model object.}
}
\value{
A data frame listing parameter names and their tags.
}
\description{
Use this function if you added the \verb{@covariates} or \verb{@input} attributes or
specified a user-defined tag (via \verb{@tag}) in one or more parameter blocks
and need to extract that information. Also, using the \verb{$INPUT} block to
declare parameters will automatically add the \verb{@input} attribute. Once these
attributes / tags are added, you can use \code{\link[=check_data_names]{check_data_names()}} to test
input data sets for names match up with parameters carrying these attributes.
}
\examples{
mod <- house()

param_tags(mod)

}
\seealso{
\code{\link[=check_data_names]{check_data_names()}}
}
