% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_derived.R
\docType{class}
\name{parameter_list-class}
\alias{parameter_list-class}
\title{S4 parameter_list class}
\description{
S4 parameter_list class
}
\details{
parameter_list is a \code{\link{numericlist-class}}
}
