% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rx.R
\name{parse_rx}
\alias{parse_rx}
\title{Parse Rx directions}
\usage{
parse_rx(x)
}
\arguments{
\item{x}{an Rx string; see \code{details} and \code{examples}}
}
\value{
An event object
}
\description{
An alternative to call int \link{ev} with
named arguments.
}
\details{
\itemize{
\item The dose is found at the start of the string by sequential digits
\item Use \code{in} to identify the dosing compartment number
\item Use \code{q} to identify the dosing interval
\item Use \code{over} to indicate an infusion and its duration
\item Use \code{x} to indicate total number of doses
\item Use \code{then} or \code{,} to separate dosing periods
\item User \code{after} to insert a lag in the start of a period
}
}
\examples{
# examples("parse_rx")

parse_rx("100")

parse_rx("100 in 2")

parse_rx("100 q12 x 3")

parse_rx("100 over 2")

parse_rx("100 q 24 x 3 then 50 q12 x 2")

parse_rx("100 then 50 q 24 after 12")

}
\seealso{
\link{ev_rx}
}
