% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsims.R
\name{plot_mrgsims}
\alias{plot_mrgsims}
\alias{plot,mrgsims,missing-method}
\alias{plot}
\alias{plot,mrgsims,formula-method}
\alias{plot,mrgsims,character-method}
\title{Generate a quick plot of simulated data}
\usage{
\S4method{plot}{mrgsims,missing}(x, limit = 16, ...)

\S4method{plot}{mrgsims,formula}(
  x,
  y,
  limit = 16,
  show.grid = TRUE,
  outer = TRUE,
  type = "l",
  lwd = 2,
  ylab = "value",
  groups = ID,
  scales = list(y = list(relation = "free")),
  fixy = NULL,
  logy = NULL,
  logbr = 0,
  ...
)

\S4method{plot}{mrgsims,character}(x, y, ...)
}
\arguments{
\item{x}{mrgsims object.}

\item{limit}{limit the the number of panels to create.}

\item{...}{other arguments passed to \code{\link[lattice:xyplot]{lattice::xyplot()}}.}

\item{y}{formula used for plotting.}

\item{show.grid}{logical indicating whether or not to draw \code{panel.grid}.}

\item{outer}{passed to \code{\link[lattice:xyplot]{lattice::xyplot()}}.}

\item{type}{passed to \code{\link[lattice:xyplot]{lattice::xyplot()}}.}

\item{lwd}{passed to \code{\link[lattice:xyplot]{lattice::xyplot()}}.}

\item{ylab}{passed to \code{\link[lattice:xyplot]{lattice::xyplot()}}.}

\item{groups}{passed to \code{\link[lattice:xyplot]{lattice::xyplot()}}.}

\item{scales}{passed to  \code{\link[lattice:xyplot]{lattice::xyplot()}}.}

\item{fixy}{make the y-axis scale the same for all variables; ignored if
\code{scales} is not a list.}

\item{logy}{plot the y variables on log scale; ignored if \code{scales} is not
a list.}

\item{logbr}{log scale breaks indicator; use \code{1} for breaks every log
unit; use \code{3} for breaks every half log unit; use \code{0} for default
breaks; ignored if \code{scales} is not a list.}
}
\description{
Generate a quick plot of simulated data
}
\examples{

mod <- mrgsolve::house(end=48, delta=0.2) \%>\% init(GUT=1000)

out <- mrgsim(mod)

plot(out)

plot(out, subset=time <= 24)

plot(out, GUT + CP ~ .)

plot(out, CP + RESP ~ time, col = "black", fixy = TRUE, lty = 2)

\dontrun{
plot(out, "CP RESP, GUT")
}

}
