% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsolve.R
\name{qsim}
\alias{qsim}
\title{Basic, simple simulation from model object}
\usage{
qsim(x, data, idata = no_idata_set(), obsonly = FALSE, tgrid = NULL,
  recsort = 1, tad = FALSE, skip_init_calc = FALSE,
  output = "mrgsims")
}
\arguments{
\item{x}{the model object}

\item{data}{can be either event object or data set}

\item{idata}{a matrix or data frame of model parameters, 
one parameter per row (see \code{\link{idata_set}})}

\item{obsonly}{if \code{TRUE}, dosing records are not included
in the output}

\item{tgrid}{a tgrid object; or a numeric vector of simulation times
or another object with an \code{stime} method}

\item{recsort}{record sorting flag.  Default value is 1.  Possible values 
are 1,2,3,4: 1 and 2 put doses in a data set after padded observations at 
the same time; 3 and 4 put those doses before padded observations at the 
same time.  2 and 4 will put doses scheduled through \code{addl} after 
observations at the same time; 1 and 3 put doses scheduled through 
\code{addl} before observations at the same time. \code{recsort} will 
not change the order of your input data set if both doses and observations 
are given.}

\item{tad}{when \code{TRUE} a column is added to simulated 
output is added showing the time since the last dose.  Only data records 
with \code{evid == 1} will be considered doses for the purposes of 
\code{tad} calculation. The \code{tad} can be properly calculated with 
a dosing lag time in the model as long as the dosing lag time (specified 
in \code{$MAIN})  is always  appropriate for any subsequent doses scheduled 
through \code{addl}.  This will always be true if the lag time doesn't 
change over time.  But it might (possibly) not hold if the lag time changes
prior to the last dose in the \code{addl} sequence.  This known limitation
shouldn't affect \code{tad} calculation in most common dosing lag time
implementations.}

\item{skip_init_calc}{don't use \code{$MAIN} to 
calculate initial conditions}

\item{output}{if \code{NULL} (the default) a mrgsims object is returned; 
otherwise, pass \code{df} to return a data.frame or \code{matrix} to 
return a matrix}
}
\description{
This is just a lighter version of \link{mrgsim}.  You
can't \code{carry_out} or \code{Request} specific columns.
}
\examples{

mod <- mrgsolve:::house()

dose <- ev(amt = 100)

out <- qsim(mod,dose)

}
