% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsim.R
\name{recmatrix}
\alias{recmatrix}
\title{Create a matrix of events for simulation.}
\usage{
recmatrix(x, times, c_indexing = TRUE)
}
\arguments{
\item{x}{an events object}

\item{times}{object that can be coerced to numeric with \code{\link{stime}}}

\item{c_indexing}{if \code{TRUE}, compartment numbers will be decremented by 1}
}
\description{
This function is for use with \code{\link{qsim}} only.
}
