% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{recover_out}
\alias{recover_out}
\title{Select items to recover into simulated output}
\usage{
recover_out(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{unquoted data set names to recover}
}
\description{
This function is similar to \code{\link[=carry_out]{carry_out()}}, in that it copies data set
items into your simulated output without having to interact with it inside
the model.  The difference is that \code{\link[=recover]{recover()}} will copy any data type:
\code{character}, \code{factors}, etc.
}
\details{
\code{recover_out} should not be used when benchmarking \code{mrgsolve}.
}
\examples{
data <- data.frame(amt = 100, CL = 5, group = "100 mg")

house() \%>\% recover_out(group) \%>\% mrgsim()

}
\seealso{
\code{\link[=carry_out]{carry_out()}}, \code{\link[=vars_out]{vars_out()}}
}
