% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\docType{methods}
\name{relocate}
\alias{relocate}
\alias{relocate,mrgmod-method}
\title{Update model or project in an model object.}
\usage{
relocate(x, ...)

\S4method{relocate}{mrgmod}(x, model = NULL, project = NULL)
}
\arguments{
\item{x}{mrgmod object}

\item{...}{passed along}

\item{model}{model name}

\item{project}{project directory}
}
\value{
updated model object
}
\description{
Update model or project in an model object.
}
