% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_cols}
\alias{rename_cols}
\title{rename columns from vector for new names}
\usage{
rename_cols(.df, new_names)
}
\arguments{
\item{.df}{dataframe to rename}

\item{new_names}{vector of names using syntax "<newname>" = "<oldname>"}
}
\description{
rename columns from vector for new names
}
\examples{
rename_cols(Theoph, c("dv" = "conc", "ID" = "Subject"))
}
