% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\docType{methods}
\name{Req}
\alias{Req}
\alias{Req,mrgmod-method}
\alias{req}
\alias{req,mrgmod-method}
\title{Request simulated output.}
\usage{
Req(x, ...)

\S4method{Req}{mrgmod}(x, ...)

req(x, ...)

\S4method{req}{mrgmod}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{unquoted names of compartments or tabled items}
}
\description{
Use this function to select, by name, either compartments or derived variables 
that have been captured (see \code{\link{CAPTURE}}).
}
\examples{
mod <- mrgsolve:::house()

mod \%>\% Req(CP,RESP) \%>\% ev(amt=1000) \%>\%  mrgsim

}

