% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{reserved}
\alias{reserved}
\title{Reserved words.}
\usage{
reserved()
}
\description{
Reserved words.
}
\details{
Note: this function is not exported; you must go into the \code{mrgsolve} namespace by using the \code{mrgsolve:::} prefix.
}
\examples{
mrgsolve:::reserved()

}
