% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/class_mrgmod.R
\name{revar}
\alias{revar}
\alias{revar,mrgmod-method}
\title{Get model random effect variances and covariances}
\usage{
revar(x, ...)

\S4method{revar}{mrgmod}(x, ...)
}
\arguments{
\item{x}{model object.}

\item{...}{passed along.}
}
\value{
A named list containing \code{omega} and \code{sigma} matrices.
}
\description{
Use this function to extract both OMEGA and SIGMA matrices from a
model object. Typical use is for display on the R console.
}
\examples{
mod <- mrgsolve::house()
revar(mod)

}
