% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modspec.R
\name{scrape_opts}
\alias{scrape_opts}
\title{Scrape options from a code block.}
\usage{
scrape_opts(x, envir = list(), def = list(), all = TRUE, marker = "=",
  narrow = TRUE)
}
\arguments{
\item{x}{data}

\item{envir}{environment from \code{$ENV}}

\item{def}{default values}

\item{all}{return all options, even those that are not in \code{def}}

\item{marker}{assignment operator; used to locate lines with options}

\item{narrow}{logical; if \code{TRUE}, only get options on lines starting with \code{>>}}
}
\value{
list with elements \code{x} (the data without options) and named options 
as specified in the block.
}
\description{
Scrape options from a code block.
}
