% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/class_mrgmod.R
\name{see}
\alias{see}
\alias{see,mrgmod-method}
\title{Print model code to the console}
\usage{
see(x, ...)

\S4method{see}{mrgmod}(x, raw = FALSE, ...)
}
\arguments{
\item{x}{model object.}

\item{...}{not used.}

\item{raw}{return the raw code.}
}
\value{
\code{NULL} is returned invisibly when \code{raw} is \code{FALSE}; when \code{raw} is
set to \code{TRUE}, the model code is returned as a character vector.
}
\description{
This is a simple way to display the model code on the R console using
the model object. The \code{raw} argument will return the model code as a
character vector.
}
