% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{show,cmt_list-method}
\alias{show,cmt_list-method}
\title{Show the compartment list}
\usage{
\S4method{show}{cmt_list}(object)
}
\arguments{
\item{object}{the object to show}
}
\description{
Show the compartment list
}
\keyword{internal}
