% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_modlist.R
\docType{methods}
\name{show,modlist-method}
\alias{show,modlist-method}
\title{Show a modlist object.}
\usage{
\S4method{show}{modlist}(object)
}
\arguments{
\item{object}{modlist object}
}
\description{
Show a modlist object.
}
