% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{show,mrgmod-method}
\alias{show,mrgmod-method}
\title{Print model details.}
\usage{
\S4method{show}{mrgmod}(object)
}
\arguments{
\item{object}{the model object}
}
\description{
Print model details.
}
