% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{show,parameter_list-method}
\alias{show,parameter_list-method}
\title{Show the parameter list}
\usage{
\S4method{show}{parameter_list}(object)
}
\arguments{
\item{object}{the object to show}
}
\description{
Show the parameter list
}
\keyword{internal}
