% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmxml.R
\name{split_nm_post}
\alias{split_nm_post}
\title{Create a list of parameters}
\usage{
split_nm_post(x, tpat = "THETA", opat = "OMEGA", spat = "SIGMA")
}
\arguments{
\item{x}{a data frame of NONMEM estimates, with one set per row}

\item{tpat}{a regular expression for detecting names in \code{x} to use for
parameters}

\item{opat}{a regular expression for detecting names in \code{x} to use for
the omega matrix}

\item{spat}{a regular expression for detecting names in \code{x} to use for
the sigma matrix}
}
\value{
An unnamed list with length equal to the number of rows in \code{x}; each
item in the list will be a list with the following data items:
\itemize{
\item \code{param}: a named list
\item \code{omega}: a matrix
\item \code{sigma}: a matrix
\item \code{row}: an integer specifying the row number
}
}
\description{
Create a list of parameters
}
