% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mrgmod.R
\name{summary.mrgmod}
\alias{summary.mrgmod}
\title{Print summary of a mrgmod object}
\usage{
\method{summary}{mrgmod}(object, ...)
}
\arguments{
\item{object}{a mrgmod object}

\item{...}{not used}
}
\description{
Print summary of a mrgmod object
}
