% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simtime.R
\docType{methods}
\name{c,tgrid-method}
\alias{c,tgrid-method}
\alias{c,tgrids-method}
\alias{tgrid_+_numeric}
\alias{+,tgrid,numeric-method}
\alias{tgrid_*_numeric}
\alias{*,tgrid,numeric-method}
\alias{tgrids_+_numeric}
\alias{+,tgrids,numeric-method}
\alias{tgrids_*_numeric}
\alias{*,tgrids,numeric-method}
\title{Operations with tgrid objects.}
\usage{
\S4method{c}{tgrid}(x, ..., recursive = FALSE)

\S4method{c}{tgrids}(x, ..., recursive = FALSE)

\S4method{+}{tgrid,numeric}(e1, e2)

\S4method{*}{tgrid,numeric}(e1, e2)

\S4method{+}{tgrids,numeric}(e1, e2)

\S4method{*}{tgrids,numeric}(e1, e2)
}
\arguments{
\item{x}{mrgmod object}

\item{...}{passed along to other methods}

\item{recursive}{not used}

\item{e1}{tgrid or tgrids object}

\item{e2}{numeric value}
}
\description{
Operations with tgrid objects.
}
