% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{touch_funs}
\alias{touch_funs}
\title{Get inits from compiled function.}
\usage{
touch_funs(x, keep_pointers = TRUE)
}
\arguments{
\item{x}{mrgmod model object}

\item{keep_pointers}{should function pointers be returned?}
}
\description{
Get inits from compiled function.
}
