% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_env}
\alias{update_env}
\title{Update objects in \code{$ENV}.}
\usage{
update_env(.x, ..., .dots = list())
}
\arguments{
\item{.x}{model object}

\item{...}{objects to update}

\item{.dots}{list of objects to updated}
}
\description{
Update objects in \code{$ENV}.
}

