% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgindata.R
\name{valid_idata_set}
\alias{valid_idata_set}
\title{Validate and prepare idata data sets for simulation}
\usage{
valid_idata_set(x, m, verbose = FALSE, quiet = FALSE)
}
\arguments{
\item{x}{data.frame or matrix}

\item{m}{a model object}

\item{verbose}{logical}

\item{quiet}{if \code{TRUE}, messages will be suppressed}
}
\value{
A numeric matrix with class \code{valid_idata_set}.
}
\description{
Validate and prepare idata data sets for simulation
}
\seealso{
\code{\link{valid_data_set}}, \code{\link{idata_set}}, 
\code{\link{data_set}}
}
