% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{vars_out}
\alias{vars_out}
\title{Select output variables}
\usage{
vars_out(x, ...)
}
\arguments{
\item{x}{a model object}

\item{...}{unquoted compartment or capture names that will appear in
simulated output}
}
\description{
This function selects compartments or capture items to appear in simulated
output. It is currently identical to \code{request}ing items via \code{\link[=Req]{Req()}} and
accomplishes the same thing as passing a character vector to the \code{Request}
argument to \code{\link[=mrgsim]{mrgsim()}} or setting \code{outvars} in \code{\link[=update]{update()}}.
}
\seealso{
\code{\link[=carry_out]{carry_out()}}, \code{\link[=recover_out]{recover_out()}}
}
