% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{wf_sweep}
\alias{wf_sweep}
\title{DEFUNCT: Simulate a sequence of parameters}
\usage{
wf_sweep(...)
}
\arguments{
\item{...}{not used.}
}
\description{
Try the 'mrgsim.sa' package instead.
}
\keyword{internal}
