% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{aboutsolver}
\alias{aboutsolver}
\title{About the DLSODA differential equation solver used by mrgsolve}
\description{
About the DLSODA differential equation solver used by mrgsolve
}
\section{DLSODA}{

\preformatted{
/*
* HISTORY:
* This is a CPP version of the LSODA library for integration into MOOSE
somulator.
* The original was aquired from
* http://www.ccl.net/cca/software/SOURCES/C/kinetics2/index.shtml and modified
by
* Heng Li <lh3lh3@gmail.com>. Heng merged several C files into one and added a
* simpler interface. [Available
here](http://lh3lh3.users.sourceforge.net/download/lsoda.c)

* The original source code came with no license or copyright
* information. Heng Li released his modification under the MIT/X11 license. I
* maintain the same license. I have removed quite a lot of text/comments from
* this library. Please refer to the standard documentation.
*
* Contact: Dilawar Singh <dilawars@ncbs.res.in>
*/
}
}

