% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{as.ev}
\alias{as.ev}
\alias{as.ev,data.frame-method}
\alias{as.ev,ev-method}
\title{Coerce an object to class ev}
\usage{
as.ev(x, ...)

\S4method{as.ev}{data.frame}(x, keep_id = TRUE, clean = FALSE, ...)

\S4method{as.ev}{ev}(x, ...)
}
\arguments{
\item{x}{An object to coerce.}

\item{...}{Not used.}

\item{keep_id}{If \code{TRUE}, \code{ID} column is retained if it exists.}

\item{clean}{If \code{TRUE}, only dosing or ID information is retained in
the result.}
}
\value{
An object with class ev.
}
\description{
Coerce an object to class ev
}
\examples{
data <- data.frame(amt = 100) 

as.ev(data)

}
