% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mrgsims.R
\name{as.list,mrgsims-method}
\alias{as.list,mrgsims-method}
\title{Coerce an mrgsims object to list}
\usage{
\S4method{as.list}{mrgsims}(x, ...)
}
\arguments{
\item{x}{an mrgsims object.}

\item{...}{not used.}
}
\description{
Coerce an mrgsims object to list
}
