% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_spec_block.R
\name{BLOCK_PARSE}
\alias{BLOCK_PARSE}
\alias{PARAM}
\alias{FIXED}
\alias{THETA}
\alias{INIT}
\alias{CMT}
\alias{CAPTURE}
\alias{HANDLEMATRIX}
\title{Functions to parse code blocks}
\usage{
PARAM(
  x,
  env,
  pos = 1,
  annotated = FALSE,
  object = NULL,
  as_object = FALSE,
  covariates = FALSE,
  ...
)

FIXED(x, env, pos = 1, annotated = FALSE, ...)

THETA(
  x,
  env,
  pos = 1,
  annotated = FALSE,
  object = NULL,
  as_object = FALSE,
  name = "THETA",
  fill = NULL,
  ...
)

INIT(x, env, pos = 1, annotated = FALSE, object = NULL, as_object = FALSE, ...)

CMT(x, env, pos = 1, annotated = FALSE, object = NULL, as_object = FALSE, ...)

CAPTURE(x, env, pos = 1, annotated = FALSE, ...)

HANDLEMATRIX(
  x,
  env,
  pos = 1,
  annotated = FALSE,
  object = NULL,
  as_object = FALSE,
  name = "...",
  type = NULL,
  oclass = "",
  prefix = "",
  labels = NULL,
  unlinked = FALSE,
  ...
)
}
\arguments{
\item{x}{data}

\item{env}{parse environment}

\item{pos}{block position}

\item{annotated}{logical}

\item{object}{the name of an object in \code{ENV}}

\item{as_object}{indicates that object code is being provided}

\item{covariates}{logical}

\item{...}{passed}

\item{name}{block name}

\item{fill}{deprecated; not used}

\item{type}{internal use}

\item{oclass}{internal use}

\item{prefix}{a prefix to add to the label}

\item{labels}{aliases to use for simulated ETA values}

\item{unlinked}{internal use}
}
\description{
Most of the basic blocks are listed in this help topic.  
But see also \code{\link{PKMODEL}} which has more-involved 
options and is documented separately.
}
\seealso{
\code{\link{PKMODEL}}
}
