% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_spec_block.R
\name{BLOCK_PARSE}
\alias{BLOCK_PARSE}
\alias{PARAM}
\alias{FIXED}
\alias{THETA}
\alias{INIT}
\alias{CMT}
\alias{CAPTURE}
\title{Functions to parse code blocks}
\usage{
PARAM(x, env, annotated = FALSE, covariates = FALSE, pos = 1,
  as_object = FALSE, ...)

FIXED(x, env, annotated = FALSE, pos = 1, ...)

THETA(x, env, annotated = FALSE, pos = 1, name = "THETA",
  fill = NULL, ...)

INIT(x, env, annotated = FALSE, pos = 1, as_object = FALSE, ...)

CMT(x, env, annotated = FALSE, pos = 1, as_object = FALSE, ...)

CAPTURE(x, env, annotated = FALSE, pos = 1, ...)
}
\arguments{
\item{x}{data}

\item{env}{parse environment}

\item{annotated}{logical}

\item{covariates}{logical}

\item{pos}{block position}

\item{as_object}{indicates that object code is being provided}

\item{...}{passed}

\item{name}{block name}

\item{fill}{data to use for block contents}
}
\description{
Most of the basic blocks are listed in this help topic.  
But see also \code{\link{PKMODEL}} which has more-involved 
options and is documented separately.
}
\seealso{
\code{\link{PKMODEL}}
}
