% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{carry.out}
\alias{carry.out}
\title{Set the \code{carry.out} argument for \code{mrgsim}.}
\usage{
carry.out(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed along}
}
\description{
Set the \code{carry.out} argument for \code{mrgsim}.
}

