% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modlib.R
\name{code}
\alias{code}
\title{Extract the code from a model.}
\usage{
code(x)
}
\arguments{
\item{x}{an mrgsolve model object}
}
\value{
a character vector of model code
}
\description{
Extract the code from a model.
}
