% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_covset.R, R/dmutate.R
\docType{methods}
\name{covset}
\alias{as.covset}
\alias{as.list,covobj-method}
\alias{as.list,covset-method}
\alias{covset}
\alias{new_covobj}
\alias{print.covobj}
\alias{print.covset}
\title{Covobj and covset objects.}
\usage{
new_covobj(x, ...)

\method{print}{covobj}(x, ...)

\S4method{as.list}{covobj}(x, ...)

\S4method{as.list}{covset}(x, ...)

\method{print}{covset}(x, ...)

covset(...)

as.covset(x)
}
\arguments{
\item{x}{a formula; may be quoted}

\item{...}{formulae to use for the covset}
}
\description{
Covobj and covset objects.

Create a set of covariates.
}

