% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{custom_tol}
\alias{custom_tol}
\alias{customize_rtol}
\alias{customize_atol}
\alias{custom_tol_list}
\alias{use_custom_tol}
\alias{use_scalar_tol}
\title{Customize tolerances for specific compartments}
\usage{
customize_rtol(.x, .rtol = list(), .default = NULL, .use = TRUE, ...)

customize_atol(.x, .atol = list(), .default = NULL, .use = TRUE, ...)

custom_tol(x)

custom_tol_list(x)

use_custom_tol(x)

use_scalar_tol(x)
}
\arguments{
\item{.x}{a model object.}

\item{.rtol}{a named numeric list or vector, where names reference
selected model compartments and relative tolerances for those
compartments.}

\item{.default}{the default tolerance value to use for compartments not
listed in \code{.rtol}, \code{.atol}, or \code{...}; if not supplied, the current scalar
value in \code{.x} will be used.}

\item{.use}{\code{logical}; if \code{TRUE}, then a call to \code{\link[=use_custom_tol]{use_custom_tol()}} will be
made prior to return; otherwise, a call to \code{\link[=use_scalar_tol]{use_scalar_tol()}} will be made;
under expected use, the value for this argument is kept \code{TRUE}, so that
whenever tolerances are customized, they will be used in the next simulation
run.}

\item{...}{\code{name}/\code{value} pairs, where \code{name} references a model compartment
and \code{value} is a new, numeric value to use for \code{rtol} or \code{atol}.}

\item{.atol}{a named numeric list or vector, where names reference
selected model compartments and absolute tolerances for those
compartments.}

\item{x}{a model object.}
}
\description{
These functions update the relative or absolute tolerance values
only for the custom tolerance configuration.

Call `use_custom_tol()` to use custom relative and absolute tolerances in 
a model; call `use_scalar_tol()` to revert to the traditional configuration
where a single `rtol` and `atol` are applied to all compartments. 
custom_tol()` returns a data frame showing tolerances for every compartment; 
`custom_tol_list()` returns the same data, but as named lists that can be 
used in other update functions.
}
\details{
New tolerance values can be supplied by either a named, numeric vector or
list via \code{.rtol} and \code{.atol} or via \code{...} or by both. If duplicate
compartment names are found in \code{...} and either \code{.rtol} or \code{.atol}, the value
passed via \code{...} will take precedence.

If customized tolerances have not been initialized yet, they will be,
assigning the current `rtol` or `atol` for every compartment. These default
values can be updated using [customize_rtol()] and [customize_atol()].
}
\examples{
mod <- house()
mod <- customize_rtol(mod, GUT = 1e-2, CENT = 1e-3)

new_tolerances <- c(GUT = 1e-4, RESP = 1e-5)
mod <- customize_rtol(mod, new_tolerances, RESP = 1e-6)

mod <- house()

mod <- use_custom_tol(mod)
mod

mod <- use_scalar_tol(mod)
mod

}
