% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{custom_tol}
\alias{custom_tol}
\alias{customize_rtol}
\alias{customize_atol}
\alias{reset_rtol}
\alias{reset_atol}
\alias{reset_tolerances}
\alias{use_custom_tol}
\alias{use_scalar_tol}
\title{Customize tolerances for specific compartments}
\usage{
customize_rtol(.x, .rtol = list(), .default = NULL, ...)

customize_atol(.x, .atol = list(), .default = NULL, ...)

reset_rtol(x, default = NULL)

reset_atol(x, default = NULL)

reset_tolerances(x, rtol = NULL, atol = NULL)

use_custom_tol(x)

use_scalar_tol(x)
}
\description{
Customize tolerances for specific compartments
}
