% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{cvec}
\alias{cvec}
\alias{cvec,character-method}
\alias{ch}
\alias{s}
\title{Create create character vectors.}
\usage{
cvec(x, ...)

\S4method{cvec}{character}(x)

ch(...)

s(...)
}
\arguments{
\item{x}{comma-separated quoted string (for \code{cvec})}

\item{...}{unquoted strings (for \code{ch})}
}
\description{
Create create character vectors.
}
\examples{

cvec("A,B,C")
ch(A,B,C)
s(A,B,C)

}
