% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsim.R
\name{data_qsim}
\alias{data_qsim}
\title{Create a matrix of events and observations for simulation}
\usage{
data_qsim(e, times)
}
\arguments{
\item{e}{an event object}

\item{times}{numeric vector of observation times or a 
\code{tgrid} object}
}
\value{
A numeric matrix with at minimum columns of 
\code{time}, \code{cmt}, \code{evid}, \code{amt}, 
\code{rate}.
}
\description{
This function is to be used with \code{\link{qsim}} only.
}
\examples{
e <- ev(amt = 100, ii = 12, addl = 2, rate = 50)

times <- tgrid(end = 240, delta = 6)

data_qsim(e, times)


}
