% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_get}
\alias{env_get}
\alias{env_get_env}
\title{Return model environment}
\usage{
env_get(x, tolist = TRUE)

env_get_env(x)
}
\arguments{
\item{x}{model object}

\item{tolist}{should the environment be coerced to \code{list}?}
}
\description{
Return model environment
}
