% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_get}
\alias{env_get}
\title{Return model environment.}
\usage{
env_get(x, tolist = TRUE)
}
\arguments{
\item{x}{model object}

\item{tolist}{should the environment be coreced to \code{list}?}
}
\description{
Return model environment.
}
