% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_ls}
\alias{env_ls}
\title{List objects in the model environment.}
\usage{
env_ls(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed to \code{\link{ls}}}
}
\description{
Each model keeps an internal environment that allows the user 
to carry any \code{R} object along.  Objects are coded in \code{$ENV}.
}
