% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ev.R
\name{mutate.ev}
\alias{mutate.ev}
\alias{select.ev}
\alias{filter.ev}
\title{dplyr verbs for event objects}
\usage{
\method{mutate}{ev}(.data, ...)

\method{select}{ev}(.data, ...)

\method{filter}{ev}(.data, ...)
}
\arguments{
\item{.data}{the event object.}

\item{...}{passed to the \code{dplyr} function.}
}
\description{
dplyr verbs for event objects
}
