% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ev.R
\docType{methods}
\name{ev_methods}
\alias{ev_methods}
\alias{nrow.ev}
\alias{names,ev-method}
\alias{as.matrix.ev}
\alias{as.data.frame.ev}
\alias{show,ev-method}
\title{Various methods for event objects}
\usage{
\method{nrow}{ev}(x)

\S4method{names}{ev}(x)

\method{as.matrix}{ev}(x, ...)

\method{as.data.frame}{ev}(x, row.names = NULL, optional = FALSE,
  add_ID = NULL, ...)

\S4method{show}{ev}(object)
}
\arguments{
\item{x}{an events object}

\item{...}{passed to various methods}

\item{row.names}{passed to \code{\link{as.data.frame}}}

\item{optional}{passed to \code{\link{as.data.frame}}}

\item{add_ID}{numeric ID of length 1 used to add \code{ID} column
only if one doesn't already exist}

\item{object}{used for \code{show}}
}
\description{
Various methods for event objects
}
