% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{ev_repeat}
\alias{ev_repeat}
\title{Repeat a block of dosing events}
\usage{
ev_repeat(x, n, wait = 0)
}
\arguments{
\item{x}{event object or dosing data frame}

\item{n}{number of times to repeat}

\item{wait}{time to wait between repeats}
}
\value{
A dosing data.frame.
}
\description{
Repeat a block of dosing events
}
