% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand.idata}
\alias{expand.idata}
\alias{expand.ev}
\alias{ev_expand}
\title{Create template data sets for simulation}
\usage{
expand.idata(...)

expand.ev(...)

ev_expand(...)
}
\arguments{
\item{...}{passed to \link{expand.grid}}
}
\description{
Create template data sets for simulation
}
\details{
An ID column is added as \code{seq(nrow(ans))} if not supplied by the user.
For \code{expand.ev}, defaults also added include \code{cmt = 1},
\code{time = 0}, \code{evid = 1}.  If \code{total} is included,
then \code{addl} is derived as \code{total} - 1. If \code{tinf} is included, then
an infusion rate is derived for row where \code{tinf} is greater than
zero.
}
\examples{
idata <- expand.idata(CL = c(1,2,3), VC = c(10,20,30))

doses <- expand.ev(amt = c(300,100), ii = c(12,24), cmt = 1)

infusion <- expand.ev(amt = 100, tinf = 2)

}
