% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand.idata}
\alias{expand.idata}
\alias{expand.ev}
\title{Create template data sets for simulation.}
\usage{
expand.idata(...)

expand.ev(...)
}
\arguments{
\item{...}{passed to \code{\link{expand.grid}}}
}
\description{
Create template data sets for simulation.
}
\details{
An ID column is added as \code{1:nrow(ans)} if not supplied by the user.  
For \code{expand.ev}, defaults
also added: \code{cmt = 1}, \code{time = 0}, \code{evid = 1}.
}
\examples{
idata <- expand.idata(CL=c(1,2,3), VC=c(10,20,30))

doses <- expand.ev(amt=c(300,100), ii=c(12,24), cmt=1)
}
