% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-tol.R
\name{get_tol}
\alias{get_tol}
\alias{get_tol_list}
\title{Extract rtol and atol information from a model object}
\usage{
get_tol(x)

get_tol_list(x)
}
\arguments{
\item{x}{a model object.}
}
\description{
Extract rtol and atol information from a model object
}
\examples{
mod <- house()
get_tol(mod)
get_tol_list(mod)

}
