% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inven.R
\name{inven}
\alias{inven}
\title{Inventory parameters in R objects.}
\usage{
inven(x, obj, need = NULL, crump = TRUE)
}
\arguments{
\item{x}{model object}

\item{obj}{the object to inventory}

\item{need}{required parameters}

\item{crump}{logical; if \code{TRUE} an error is generated when any parameter
in \code{need} is not found in the object}
}
\description{
Inventory parameters in R objects.
}
