% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mrgmod.R
\name{is.mrgmod}
\alias{is.mrgmod}
\title{Check if an object is a model object.}
\usage{
is.mrgmod(x)
}
\arguments{
\item{x}{any object}
}
\value{
\code{TRUE} if \code{x} inherits \code{mrgsims}.
}
\description{
The function checks to see if the object is either
 \code{mrgmod} or \code{packmod}.
}
