% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set.R
\name{lctran}
\alias{lctran}
\title{Convert select upper case column names to lower case to conform to mrgsolve data expectations.}
\usage{
lctran(data)
}
\arguments{
\item{data}{an nmtran-like data frame}
}
\value{
A data.frame with renamed columns.
}
\description{
Convert select upper case column names to lower case to conform to mrgsolve data expectations.
}
\details{
Columns that will be renamed with lower case versions: \code{AMT}, \code{II}, \code{SS}, \code{CMT}, \code{ADDL}, \code{RATE}, \code{EVID}, \code{TIME}.  If a lower case version
of these names exist in the data set, the column will not be renamed.
}
