% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/class_mrgmod.R
\docType{methods}
\name{loadso}
\alias{loadso}
\alias{loadso,mrgmod-method}
\title{Load the model shared object.}
\usage{
loadso(x, ...)

\S4method{loadso}{mrgmod}(x, ...)
}
\arguments{
\item{x}{the model object}

\item{...}{passed along}
}
\description{
Load the model shared object.
}
