% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mrgmod.R
\name{ls_env}
\alias{ls_env}
\title{List objects in \code{$ENV}.}
\usage{
ls_env(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed to \code{\link{ls}}}
}
\description{
List objects in \code{$ENV}.
}

