% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlist.R
\name{matlist}
\alias{matlist}
\alias{zero.re}
\alias{zero.re,mrgmod-method}
\alias{zero_re}
\alias{as.list,matlist-method}
\alias{as.matrix,matlist-method}
\alias{names.matlist}
\alias{length.matlist}
\alias{labels,matlist-method}
\alias{dim,matlist-method}
\alias{nrow,matlist-method}
\alias{show,matlist-method}
\title{Methods for working with matrix-list objects}
\usage{
zero.re(.x, ...)

\S4method{zero.re}{mrgmod}(.x, ..., .drop = FALSE)

zero_re(...)

\S4method{as.list}{matlist}(x, ...)

\S4method{as.matrix}{matlist}(x, ...)

\method{names}{matlist}(x)

\method{length}{matlist}(x)

\S4method{labels}{matlist}(object, ...)

\S4method{dim}{matlist}(x)

\S4method{nrow}{matlist}(x)

\S4method{show}{matlist}(object)
}
\arguments{
\item{.x}{a matlist object}

\item{...}{passed along}

\item{.drop}{if \code{TRUE}, \code{zero_re} will drop \code{omega}
or \code{sigma} or both entirely}

\item{x}{a matlist object}

\item{object}{passed to showmatlist}
}
\description{
Methods for working with matrix-list objects
}
\keyword{internal}
