% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlist.R
\name{matlist}
\alias{matlist}
\alias{as.list,matlist-method}
\alias{as.matrix,matlist-method}
\alias{names.matlist}
\alias{length.matlist}
\alias{labels,matlist-method}
\alias{dim,matlist-method}
\alias{nrow,matlist-method}
\alias{show,matlist-method}
\title{Methods for working with matrix-list objects}
\usage{
\S4method{as.list}{matlist}(x, detailed = FALSE, ...)

\S4method{as.matrix}{matlist}(x, ...)

\method{names}{matlist}(x)

\method{length}{matlist}(x)

\S4method{labels}{matlist}(object, ...)

\S4method{dim}{matlist}(x)

\S4method{nrow}{matlist}(x)

\S4method{show}{matlist}(object)
}
\arguments{
\item{x}{a matlist object}

\item{detailed}{if \code{TRUE}, then a simple list of matrices is returned;
otherwise, then entire \code{matlist} object data is returned along with the
name of the \code{class} (e.g. either \code{omegalist} or \code{sigmalist}) as well
as the \code{names} of the matrices}

\item{...}{passed through to other methods}

\item{object}{passed to showmatlist}

\item{.x}{a matlist object}
}
\description{
Methods for working with matrix-list objects
}
\keyword{internal}
