% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modspec.R
\name{modelparse}
\alias{modelparse}
\alias{modelparse_rmd}
\title{Parse model specification text}
\usage{
modelparse(
  txt,
  split = FALSE,
  drop_blank = TRUE,
  comment_re = c("//", "##"),
  keep_mapping = FALSE
)

modelparse_rmd(txt, split = FALSE, drop_blank = TRUE, comment_re = "//")
}
\arguments{
\item{txt}{model specification text.}

\item{split}{logical; if \code{TRUE}, \code{txt} will be split on \verb{\\n} before
processing.}

\item{drop_blank}{logical; \code{TRUE} if blank lines wilil be dropped.}

\item{comment_re}{regular expression to identify comments.}

\item{keep_mapping}{if \code{TRUE}, parse information will be retained as
attributes on the parsed model code.}
}
\description{
Parse model specification text
}
\examples{
file <- file.path(modlib(), "pk1.cpp")

code <- readLines(file)

modelparse(code)

}
\seealso{
\code{\link[=modelsplit]{modelsplit()}} and \code{\link[=modelunsplit]{modelunsplit()}} for a non-destructive
split/reassemble alternative.
}
\keyword{internal}
