% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modspec.R
\name{modelparse}
\alias{modelparse}
\alias{modelparse_rmd}
\title{Parse model specification text}
\usage{
modelparse(txt, split = FALSE, drop_blank = TRUE, comment_re = c("//", "##"))

modelparse_rmd(txt, split = FALSE, drop_blank = TRUE, comment_re = "//")
}
\arguments{
\item{txt}{model specification text}

\item{split}{logical}

\item{drop_blank}{logical; \code{TRUE} if blank lines are to be dropped}

\item{comment_re}{regular expression for comments}
}
\description{
Parse model specification text
}
\examples{
file <- file.path(modlib(), "pk1.cpp")

modelparse(readLines(file))

}
\keyword{internal}
