% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsims.R
\name{mrgsims_dplyr}
\alias{mrgsims_dplyr}
\alias{as.tbl.mrgsims}
\alias{pull.mrgsims}
\alias{filter.mrgsims}
\alias{filter_sims}
\alias{group_by.mrgsims}
\alias{distinct.mrgsims}
\alias{mutate.mrgsims}
\alias{mutate_sims}
\alias{summarise.each}
\alias{summarise.mrgsims}
\alias{do.mrgsims}
\alias{select.mrgsims}
\alias{select_sims}
\alias{slice.mrgsims}
\alias{as_data_frame.mrgsims}
\alias{as_tibble.mrgsims}
\title{Methods for handling output with dplyr verbs}
\usage{
\method{as.tbl}{mrgsims}(x, ...)

\method{pull}{mrgsims}(.data, ...)

\method{filter}{mrgsims}(.data, ...)

filter_sims(.data, ...)

\method{group_by}{mrgsims}(.data, ..., add = FALSE)

\method{distinct}{mrgsims}(.data, ..., .keep_all = FALSE)

\method{mutate}{mrgsims}(.data, ...)

mutate_sims(.data, ...)

\method{summarise}{each}(.data, funs, ...)

\method{summarise}{mrgsims}(.data, ...)

\method{do}{mrgsims}(.data, ..., .dots)

\method{select}{mrgsims}(.data, ...)

select_sims(.data, ...)

\method{slice}{mrgsims}(.data, ...)

as_data_frame.mrgsims(.data_, ...)

\method{as_tibble}{mrgsims}(.data_, ...)
}
\arguments{
\item{x}{mrgsims object}

\item{...}{passed to other methods}

\item{.data}{passed to various \code{dplyr} functions}

\item{add}{passed to \code{dplyr::group_by}}

\item{.keep_all}{passed to \code{dplyr::distinct}}

\item{funs}{passed to \code{dplyr::summarise_each}}

\item{.dots}{passed to various \code{dplyr} functions}

\item{.data_}{mrgsims object}
}
\description{
Methods for handling output with dplyr verbs
}
\details{
For the \code{select_sims} function, the dots \code{...} must be either 
compartment names or variables in \code{$CAPTURE}.  An error will be
generated if no valid names are selected or the names for selection are 
not found in the simulated output.
}
