% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsims.R
\name{mrgsims_dplyr}
\alias{mrgsims_dplyr}
\alias{as.tbl.mrgsims}
\alias{filter_.mrgsims}
\alias{group_by_.mrgsims}
\alias{distinct_.mrgsims}
\alias{mutate_.mrgsims}
\alias{summarise.each}
\alias{summarise_.mrgsims}
\alias{do_.mrgsims}
\alias{select_.mrgsims}
\alias{slice_.mrgsims}
\alias{as_data_frame.mrgsims}
\title{Methods for handling output with dplyr verbs.}
\usage{
\method{as.tbl}{mrgsims}(x, ...)

\method{filter_}{mrgsims}(.data, ..., .dots)

\method{group_by_}{mrgsims}(.data, ..., .dots, add = FALSE)

\method{distinct_}{mrgsims}(.data, ..., .dots, .keep_all = FALSE)

\method{mutate_}{mrgsims}(.data, ..., .dots)

summarise.each(.data, funs, ...)

\method{summarise_}{mrgsims}(.data, ..., .dots)

\method{do_}{mrgsims}(.data, ..., .dots)

\method{select_}{mrgsims}(.data, ..., .dots)

\method{slice_}{mrgsims}(.data, ...)

\method{as_data_frame}{mrgsims}(.data_, ...)
}
\arguments{
\item{x}{mrgsims object}

\item{...}{passed to other methods}

\item{.data}{passed to various \code{dplyr} functions}

\item{.dots}{passed to various \code{dplyr} functions}

\item{add}{passed to \code{dplyr::group_by_}}

\item{.keep_all}{passed to \code{dplyr::distinct_}}

\item{funs}{passed to \code{dplyr::summarise_each}}

\item{.data_}{mrgsims object}
}
\description{
Methods for handling output with dplyr verbs.
}
