% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{obsonly}
\alias{obsonly}
\title{Collect only observation records in the simulated output}
\usage{
obsonly(x, value = TRUE, ...)
}
\arguments{
\item{x}{model object.}

\item{value}{the value for `obsonly`.}

\item{...}{not used.}
}
\description{
Collect only observation records in the simulated output
}
\details{
There is also a `obsonly` argument to [mrgsim()] that can 
be set to accomplish the same thing as a call to `obsonly()` in 
the pipeline.
}
