% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{obsonly}
\alias{obsonly}
\title{Collect only observations in the simulated output.}
\usage{
obsonly(x, value = TRUE, ...)
}
\arguments{
\item{x}{model object}

\item{value}{the value for \code{obsonly}}

\item{...}{passed along}
}
\description{
Collect only observations in the simulated output.
}
\details{
There is also a \code{obsonly} argument to \code{\link{mrgsim}} that can 
be set to accomplish the same thing as a call to \code{obsonly} in 
the pipeline.
}
