% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mrgmod.R
\name{outvars}
\alias{outvars}
\title{Show names of current output variables}
\usage{
outvars(x, unlist = FALSE)
}
\arguments{
\item{x}{model object.}

\item{unlist}{if \code{TRUE} then a character vector (rather than list) is
returned.}
}
\value{
When \code{unlist} is \code{FALSE} (default) : a named list, with \code{cmt} showing names
of output compartments and \code{capture} giving names of output variables in
capture.  When \code{unlist} is \code{TRUE}, then a single, unnamed character vector
of outvar names is returned.
}
\description{
Outputs can include model compartments or variables defined in the model
that have been marked to \code{capture} in simulated output.
}
\examples{
mod <- mrgsolve::house()
outvars(mod)

}
