% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inven.R
\name{param_tags}
\alias{param_tags}
\title{Return parameter attributes and tags}
\usage{
param_tags(x)
}
\arguments{
\item{x}{mrgsolve model object.}
}
\value{
A data frame listing paramter names and their tags.
}
\description{
Use this function if you added attributes to the parameters in the model
specification and need to extract that information. Attributes include
\verb{@covariates} and \verb{@input}; paramters can also be assigned an arbitrary
tag with the \verb{@tag} attribute. Using the \verb{$INPUT} block to declare
parameters will automatically add the \verb{@input} attribute. Once these
attributes are added, you can use \code{\link[=check_data_names]{check_data_names()}} to test input
data sets for names match up with paramters carrying these attributes.
}
\examples{
mod <- house()

param_tags(mod)

}
\seealso{
\code{\link[=check_data_names]{check_data_names()}}
}
