% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsim.R
\name{qsim}
\alias{qsim}
\alias{qsim_df}
\title{DEPRECATED: A quick simulation function}
\usage{
qsim(x, e = NULL, idata = NULL, req = NULL, tgrid = NULL,
  skip_init_calc = FALSE)

qsim_df(...)
}
\arguments{
\item{x}{model object}

\item{e}{event object}

\item{idata}{individual data set}

\item{req}{compartments to request}

\item{tgrid}{\code{tgrid} object; used if \code{e} is an \code{ev} object}

\item{skip_init_calc}{not used}

\item{...}{passed to \code{\link{qsim}}}
}
\description{
DEPRECATED: A quick simulation function
}
\keyword{internal}
