% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{realize_addl}
\alias{realize_addl}
\alias{realize_addl.data.frame}
\alias{realize_addl.ev}
\title{Make addl doses explicit in an event object or data set.}
\usage{
realize_addl(x, ...)

\method{realize_addl}{data.frame}(x, ...)

\method{realize_addl}{ev}(x, ...)
}
\arguments{
\item{x}{a \code{data_set} data frame or an \code{ev} object (see details)}

\item{...}{not used}
}
\description{
Make addl doses explicit in an event object or data set.
}
\details{
Required data elements: \code{addl} and \code{ii}.
}
