% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{reset_tol}
\alias{reset_tol}
\alias{reset_tolerances}
\alias{reset_rtol}
\alias{reset_atol}
\title{Reset all model tolerances}
\usage{
reset_tolerances(x, rtol = NULL, atol = NULL)

reset_rtol(x, rtol = NULL)

reset_atol(x, atol = NULL)
}
\arguments{
\item{x}{a model object.}

\item{rtol}{global relative tolerance for both scalar and customized
configurations.}

\item{atol}{global absolute tolerance for both scalar and customized
configurations.}
}
\description{
These functions reset both scalar and customized values for both 
relative and absolute tolerances. The functions do _not_ change 
the which tolerance configuration is used for simulation (e.g., 
scalar or customized); see [use_custom_tol()] and [use_scalar_tol()]
to make that change in the model object.
}
\examples{
mod <- house()
mod <- reset_tolerances(mod, rtol = 1e-6, atol = 1e-10)
mod

}
