% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modspec.R
\name{scrape_and_call}
\alias{scrape_and_call}
\title{Scrape options and pass to function.}
\usage{
scrape_and_call(x, env, pass, ...)
}
\arguments{
\item{x}{data}

\item{env}{parse environment}

\item{pass}{function to call}

\item{...}{dots}
}
\description{
Scrape options and pass to function.
}
\details{
Attributes of \code{x} are also scraped and merged with options.
}
