% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/class_mrgmod.R
\docType{methods}
\name{see}
\alias{see}
\alias{see,mrgmod-method}
\title{Print model code to the console.}
\usage{
see(x, ...)

\S4method{see}{mrgmod}(x, raw = FALSE, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed along}

\item{raw}{return the raw code}
}
\value{
invisible NULL
}
\description{
Print model code to the console.
}
