% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simargs}
\alias{simargs}
\alias{simargs.mrgmod}
\title{Access or clear arguments for calls to mrgsim.}
\usage{
simargs(x, ...)

\method{simargs}{mrgmod}(x, clear = FALSE, ...)
}
\arguments{
\item{x}{model object}

\item{...}{passed along}

\item{clear}{logical indicating whether or not clear args from the model object}
}
\value{
If \code{clear} is \code{TRUE}, the argument list is cleared and the model object is returned.  Otherwise, the argument list is returned.
}
\description{
Access or clear arguments for calls to mrgsim.
}
\examples{
mod <- mrgsolve:::house()
mod \%>\% Req(CP,RESP) \%>\% carry_out(evid,WT,FLAG) \%>\% simargs

}
