% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mrgmod.R
\name{soloc}
\alias{soloc}
\title{Return the location of the model shared object.}
\usage{
soloc(x, short = FALSE)
}
\arguments{
\item{x}{model object}

\item{short}{logical; if \code{TRUE}, \code{soloc} will be rendered  with a short path name}
}
\description{
Return the location of the model shared object.
}
\examples{
mod <- mrgsolve:::house()
soloc(mod)

}
