% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{use_custom_tol}
\alias{use_custom_tol}
\alias{use_scalar_tol}
\title{Update a model object to use custom or scalar tolerances}
\usage{
use_custom_tol(x)

use_scalar_tol(x)
}
\arguments{
\item{x}{a model object.}
}
\description{
Update a model object to use custom or scalar tolerances
}
\examples{
mod <- house()

mod <- use_custom_tol(mod)
mod

mod <- use_scalar_tol(mod)
mod

}
