% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modspec.R
\name{warn_integer_division}
\alias{warn_integer_division}
\title{Warn for literal integer division}
\usage{
warn_integer_division(x)
}
\arguments{
\item{x}{string to check.}
}
\description{
Warn for literal integer division
}
\examples{
warn_integer_division("THETA(1) * pow(WT/70, 3/4)")
warn_integer_division("3.0/4")

}
