% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{wf_sweep}
\alias{wf_sweep}
\title{Simulate a sequence of parameters}
\usage{
wf_sweep(mod, ..., n = 10, cv = 30, nsd = 2)
}
\arguments{
\item{mod}{a model object}

\item{...}{unquoted parameter names}

\item{n}{number of values to simulate}

\item{cv}{coefficient of variation; used to determine the minimum and
maximum values for the sweep}

\item{nsd}{the number of standard deviations over which to sweep}
}
\value{
A mrgsims object with a character columns named \code{.name}
indicating the parameter name and \code{.value} indicating the sweep
value.
}
\description{
This function is experimental and may change or go away at any time without
notice.
}
\examples{
mod <- mrgsolve::house() \%>\% zero_re()

mod \%>\% 
  ev(amt = 100) \%>\% 
  wf_sweep(CL, VC) \%>\% 
  plot(CP~time|.name)

}
\keyword{internal}
